+++ examples/Icon/InventoryIcons.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+import Icon from 'src/Icon';
+
+function InventoryIconsSet() {
+    return (
+        <div>
+            <Icon type="inventory-category" />
+            <Icon type="inventory-item" />
+        </div>
+    );
+}
+
+export default InventoryIconsSet;