+++ CHANGELOG.md
@@ -7,6 +7,27 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 
+
+## [0.9.0]
+### Added
+- Add `fontello.config.json` at project root.
+- Add inventory icons(`iventory-category` and `inventory-item`).
+- Add [@kadira/storybook](https://github.com/storybooks/storybook/) to replace `demo/`.
+- Create `examples/` folder and migrate component docs to stories.
+- Create `addPropsTable()` custom addon to show component's propTypes in table, which modified from `addon-info`.
+- Add `storybook` env in babelrc, which included cjs transpiler. 
+- Deploy built storybook to `gh-pages` branch with running `npm run ghpages`. (#35)
+- Add [demo link](http://ichef.github.io/gypcrete) in README.
+
+### Changed
+- Ignore `public/` path in eslint.
+
+### Removed
+- Remove `demo/` folder, all examples migrated to `examples/`.
+- Remove webpack base & demo config, only leave `config/webpack.prod`.
+- Remove demo related dev dependencies and `react-hot-loader`.
+
+
 ## [0.8.1]
 ### Changed
 - Vertical padding for row components are now `8px` instead of `4px`.